server done?

+++ Keepers/Services/VaultKeepsService.cs
@@ -1,3 +1,4 @@
+using System;
 using Keepers.Models;
 using Keepers.Repositories;
 
@@ -16,16 +17,22 @@ namespace Keepers.Services
 
     internal VaultKeep Create(VaultKeep newVaultKeep, Account user)
     {
-      // if (newVaultKeep.CreatorId != user.Id)
-      // {
-      //   throw new Exception("You cannot add this keep to another user's vault");
-      // }
       VaultKeep vaultKeep = _vkRepo.Create(newVaultKeep);
+      Vault vault = _vaultsService.GetOne(newVaultKeep.VaultId, user.Id);
+      if (vaultKeep.CreatorId != user.Id)
+      {
+        throw new Exception("You cannot add this keep to another user's vault");
+      }
       return vaultKeep;
     }
 
     internal string Delete(int id, string userId)
     {
+      Vault vault = _vaultsService.GetOne(id, userId);
+      if (vault.CreatorId != userId)
+      {
+        throw new Exception("You cannot delete a keep that belongs to another user's vault");
+      }
       _vkRepo.Delete(id);
       return $"This vaultkeep has been deleted";
     }
